Reran chains and updated analysis notebooks for ssfr-dependent model.
The best-fit model overproduces bulges in high-mass galaxies and under-produces bulges in low-mass galaxies. As M* increases, the bulge-dominated fraction gets ever-larger in the best-fit model, whereas in the data, this fraction levels off around 10**11. In zero-disruption models, for M*>10**11 the bulge-dominated fraction is just 0.1, whereas SDSS it's 0.3. So from this we learn that modifying the treatment of mergers (diverting some merging mass into the disk, for example) will not remedy the problem. To remedy the problem, we need to change the shape of F_bulge(M*) to raise it at low-mass and lower it at high-mass.